python project/dumping/parser_kaspi/parser_kaspi/settings/dev.py
import os

from .base import *
from dotenv import load_dotenv

load_dotenv()

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = False
DEFAULT_FILE_STORAGE = 'django.core.files.storage.FileSystemStorage'

MEDIA_URL = '/'
MEDIA_ROOT = os.path.join(BASE_DIR, '..')
