python project/dumping/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==4.2.0
boto3==1.34.144
botocore==1.34.144
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==4.2.11
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==24.1
django-import-export==3.3.7
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
ipython==8.12.3
jedi==0.19.1
jmespath==1.0.1
kombu==5.3.5
magic-filter==1.0.12
MarkupPy==1.14
matplotlib-inline==0.1.7
middleware==1.2.3
multidict==6.0.5
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.31.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
telegram==0.0.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
django-redis==5.4.0
whitenoise==6.7.0
jinja2~=3.1.4